minor bug fix

+++ train_test_split.py
@@ -57,7 +57,7 @@ fit_dic_P = {
     "costh": data_fit_P["costh"],
 }
 
-events_D1 = events_D[(events_D.true_mass > 4.5) & (events_D.occuD1 < 300.)].to_numpy()[:, data_fit_P["mass"].shape[0]* 0.08]
+events_D1 = events_D[(events_D.true_mass > 4.5) & (events_D.occuD1 < 300.)].to_numpy()[:, int(data_fit_P["mass"].shape[0]* 0.08)]
 
 data_train_val_D, data_test_fit_D = train_test_split(events_D1, test_size=0.5, shuffle=True)
 data_train_D, data_val_D = train_test_split(data_train_val_D, test_size=0.5, shuffle=True)
